Expected probability of a successful program for bias adjustment programs with normally distributed outcomes

To discount for overoptimistic results in phase II when calculating the optimal sample size in phase III, it is necessary to use the following functions, which each describe a specific case:

EPsProg_normal_L(): calculates the expected probability of a successful for an additive adjustment factor (i.e adjust the lower bound of the one-sided confidence interval), however the go-decision is not affected by the bias adjustment

EPsProg_normal_L2(): calculates the expected probability of a successful for an additive adjustment factor (i.e adjust the lower bound of the one-sided confidence interval) when the go-decision is also affected by the bias adjustment

EPsProg_normal_R(): calculates the expected probability of a successful for a multiplicative adjustment factor (i.e. use estimate with a retention factor), however the go-decision is not affected by the bias adjustment

EPsProg_normal_R2(): calculates the expected probability of a successful for a multiplicative adjustment factor (i.e. use estimate with a retention factor) when the go-decision is also affected by the bias adjustment

Usage

EPsProg_normal_L(
  kappa,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  fixed
)

EPsProg_normal_L2(
  kappa,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  fixed
)

EPsProg_normal_R(
  kappa,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  fixed
)

EPsProg_normal_R2(
  kappa,
  n2,
  Adj,
  alpha,
  beta,
  step1,
  step2,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  fixed
)

Arguments

kappa

threshold value for the go/no-go decision rule

n2

total sample size for phase II; must be even number

Adj

adjustment parameter

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

step1

lower boundary for effect size

step2

upper boundary for effect size

w

weight for mixture prior distribution

Delta1

assumed true treatment effect for standardized difference in means

Delta2

assumed true treatment effect for standardized difference in means

in1

amount of information for Delta1 in terms of sample size

in2

amount of information for Delta2 in terms of sample size

a

lower boundary for the truncation

b

upper boundary for the truncation

fixed

choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect

Value

The output of the the functions EPsProg_normal_L(), EPsProg_normal_L2(), EPsProg_normal_R() and EPsProg_normal_R2() is the expected probability of a successful program.

Last Edited By

Johannes Cepicka

Last Edit Date

2022-04-23

Examples

res <- EPsProg_normal_L(kappa = 0.1, n2 = 50, Adj = 0, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 step1 = 0, step2 = 0.5,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, fixed = FALSE)
          res <- EPsProg_normal_L2(kappa = 0.1, n2 = 50, Adj = 0, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 step1 = 0, step2 = 0.5,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, fixed = FALSE)
          res <- EPsProg_normal_R(kappa = 0.1, n2 = 50, Adj = 1, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 step1 = 0, step2 = 0.5,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, fixed = FALSE)
          res <- EPsProg_normal_R2(kappa = 0.1, n2 = 50, Adj = 1, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 step1 = 0, step2 = 0.5,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, fixed = FALSE)