Expected probability of a successful program for multiple endpoints and normally distributed outcomes

This function calculates the probability that our drug development program is successful. Successful is defined as both endpoints showing a statistically significant positive treatment effect in phase III.

Usage

EPsProg_multiple_normal(
  kappa,
  n2,
  alpha,
  beta,
  Delta1,
  Delta2,
  sigma1,
  sigma2,
  step11,
  step12,
  step21,
  step22,
  in1,
  in2,
  fixed,
  rho,
  rsamp
)

Arguments

kappa

threshold value for the go/no-go decision rule;

n2

total sample size for phase II; must be even number

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

Delta1

assumed true treatment effect given as difference in means for endpoint 1

Delta2

assumed true treatment effect given as difference in means for endpoint 2

sigma1

standard deviation of first endpoint

sigma2

standard deviation of second endpoint

step11

lower boundary for effect size for first endpoint

step12

lower boundary for effect size for second endpoint

step21

upper boundary for effect size for first endpoint

step22

upper boundary for effect size for second endpoint

in1

amount of information for Delta1 in terms of sample size

in2

amount of information for Delta2 in terms of sample size

fixed

choose if true treatment effects are fixed or random, if TRUE then Delta1 is used as fixed effect

rho

correlation between the two endpoints

rsamp

sample data set for Monte Carlo integration

Value

The output of the function EPsProg_multiple_normal() is the expected probability of a successfull program, when going to phase III.