Expected probability of a successful program for multitrial programs in a time-to-event setting

These functions calculate the expected probability of a successful program given the parameters. Each function represents a specific strategy, e.g. the function EpsProg3() calculates the expected probability if three phase III trials are performed. The parameter case specifies how many of the trials have to be successful, i.e. how many trials show a significantly relevant positive treatment effect.

Usage

EPsProg2(HRgo, d2, alpha, beta, w, hr1, hr2, id1, id2, case, size, fixed)

EPsProg3(HRgo, d2, alpha, beta, w, hr1, hr2, id1, id2, case, size, fixed)

EPsProg4(HRgo, d2, alpha, beta, w, hr1, hr2, id1, id2, case, size, fixed)

Arguments

HRgo

threshold value for the go/no-go decision rule

d2

total number of events in phase II

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

w

weight for mixture prior distribution

hr1

first assumed true treatment effect on HR scale for prior distribution

hr2

second assumed true treatment effect on HR scale for prior distribution

id1

amount of information for hr1 in terms of number of events

id2

amount of information for hr2 in terms of number of events

case

choose case: "at least 1, 2 or 3 significant trials needed for approval"

size

size category "small", "medium" or "large"

fixed

choose if true treatment effects are fixed or random

Value

The output of the function EPsProg2(), EPsProg3() and EPsProg4() is the expected probability of a successful program when performing several phase III trials (2, 3 or 4 respectively)

Details

The following cases can be investigated by the software:

Two phase III trials

Case 1: Strategy 1/2; at least one trial significant, the treatment effect of the other one at least showing in the same direction

Case 2: Strategy 2/2; both trials significant

Three phase III trials

Case 2: Strategy 2/3; at least two trials significant, the treatment effect of the other one at least showing in the same direction

Case 3: Strategy 3/3; all trials significant

Four phase III trials

Case 3: Strategy 3/4; at least three trials significant, the treatment effect of the other one at least showing in the same direction

Examples

EPsProg2(HRgo = 0.8, d2 = 50,  alpha = 0.025, beta = 0.1, 
                                 w = 0.3, hr1 =  0.69, hr2 = 0.81, 
                                 id1 = 210, id2 = 420, case = 2, size = "small",
                                 fixed = FALSE)
#> [1] 0.03607598
          EPsProg3(HRgo = 0.8, d2 = 50,  alpha = 0.025, beta = 0.1, 
                                 w = 0.3, hr1 =  0.69, hr2 = 0.81, 
                                 id1 = 210, id2 = 420, case = 2, size = "small",
                                 fixed = TRUE)
#> [1] 0.02454443
          EPsProg4(HRgo = 0.8, d2 = 50,  alpha = 0.025, beta = 0.1, 
                                 w = 0.3, hr1 =  0.69, hr2 = 0.81, 
                                 id1 = 210, id2 = 420, case = 3, size = "small",
                                 fixed = TRUE)
#> [1] 0.003566271