Expected sample size for phase III for bias adjustment programs and binary distributed outcomes

To discount for overoptimistic results in phase II when calculating the optimal sample size in phase III, it is necessary to use the functions En3_binary_L(), En3_binary_L2(), En3_binary_R() and En3_binary_R2(). Each function describes a specific case:

En3_binary_L(): calculates the optimal sample size for an additive adjustment factor (i.e adjust the lower bound of the one-sided confidence interval), however the go-decision is not affected by the bias adjustment

En3_binary_L2(): calculates the optimal sample size for an additive adjustment factor (i.e adjust the lower bound of the one-sided confidence interval) when the go-decision is also affected by the bias adjustment

En3_binary_R(): calculates the optimal sample size for a multiplicative adjustment factor (i.e. use estimate with a retention factor), however the go-decision is not affected by the bias adjustment

En3_binary_R2(): calculates the optimal sample size for a multiplicative adjustment factor (i.e. use estimate with a retention factor) when the go-decision is also affected by the bias adjustment

Usage

En3_binary_L(RRgo, n2, Adj, alpha, beta, p0, w, p11, p12, in1, in2, fixed)

En3_binary_L2(RRgo, n2, Adj, alpha, beta, p0, w, p11, p12, in1, in2, fixed)

En3_binary_R(RRgo, n2, Adj, alpha, beta, p0, w, p11, p12, in1, in2, fixed)

En3_binary_R2(RRgo, n2, Adj, alpha, beta, p0, w, p11, p12, in1, in2, fixed)

Arguments

RRgo

threshold value for the go/no-go decision rule

n2

total sample size for phase II; must be even number

Adj

adjustment parameter

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

p0

assumed true rate of control group

w

weight for mixture prior distribution

p11

assumed true rate of treatment group

p12

assumed true rate of treatment group

in1

amount of information for p11 in terms of sample size

in2

amount of information for p12 in terms of sample size

fixed

choose if true treatment effects are fixed or random, if TRUE p11 is used as fixed effect

Value

The output of the the functions En3_binary_L, En3_binary_L2, En3_binary_R and En3_binary_R2 is the expected number of participants in phase III.

Examples

res <- En3_binary_L(RRgo = 0.8, n2 = 50, Adj = 0, 
                              alpha = 0.025, beta = 0.1, p0 = 0.6,  w = 0.3,
                              p11 =  0.3, p12 = 0.5, in1 = 300, in2 = 600, 
                              fixed = FALSE)
          res <-  En3_binary_L2(RRgo = 0.8, n2 = 50, Adj = 0, 
                              alpha = 0.025, beta = 0.1, p0 = 0.6,  w = 0.3,
                              p11 =  0.3, p12 = 0.5, in1 = 300, in2 = 600, 
                              fixed = FALSE)
          res <- En3_binary_R(RRgo = 0.8, n2 = 50, Adj = 1, 
                              alpha = 0.025, beta = 0.1, p0 = 0.6,  w = 0.3,
                              p11 =  0.3, p12 = 0.5, in1 = 300, in2 = 600, 
                              fixed = FALSE)
          res <- En3_binary_R2(RRgo = 0.8, n2 = 50, Adj = 1, 
                              alpha = 0.025, beta = 0.1, p0 = 0.6,  w = 0.3,
                              p11 =  0.3, p12 = 0.5, in1 = 300, in2 = 600, 
                              fixed = FALSE)