Expected probability to go to phase III for bias adjustment programs with time-to-event outcomes

In the case we do not only want do discount for overoptimistic results in phase II when calculating the sample size in phase III, but also when deciding whether to go to phase III or not the functions Epgo_L2 and Epgo_R2 are necessary. The function Epgo_L2 uses an additive adjustment parameter (i.e. adjust the lower bound of the one-sided confidence interval), the function Epgo_R2 uses a multiplicative adjustment parameter (i.e. use estimate with a retention factor)

Usage

Epgo_L2(HRgo, d2, Adj, w, hr1, hr2, id1, id2, fixed)

Epgo_R2(HRgo, d2, Adj, w, hr1, hr2, id1, id2, fixed)

Arguments

HRgo

threshold value for the go/no-go decision rule

d2

total number of events for phase II; must be even number

Adj

adjustment parameter

w

weight for mixture prior distribution

hr1

first assumed true treatment effect on HR scale for prior distribution

hr2

second assumed true treatment effect on HR scale for prior distribution

id1

amount of information for hr1 in terms of number of events

id2

amount of information for hr2 in terms of number of events

fixed

choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect

Value

The output of the functions Epgo_L2 and Epgo_R2 is the expected probability to go to phase III with conservative decision rule and sample size calculation.

Examples

res <- Epgo_L2(HRgo = 0.8, d2 = 50, Adj = 0.4,  
                                w = 0.3, hr1 = 0.69, hr2 = 0.81, 
                                id1 = 280, id2 = 420, fixed = FALSE)
          res <- Epgo_R2(HRgo = 0.8, d2 = 50, Adj = 0.9,  
                                w = 0.3, hr1 = 0.69, hr2 = 0.81, 
                                id1 = 280, id2 = 420, fixed = FALSE)