Expected sample size for phase III for multiarm programs with normally distributed outcomes

Given phase II results are promising enough to get the "go"-decision to go to phase III this function now calculates the expected sample size for phase III given the cases and strategies listed below. The results of this function are necessary for calculating the utility of the program, which is then in a further step maximized by the optimal_multiarm_normal() function

Usage

Ess_normal(kappa, n2, alpha, beta, Delta1, Delta2, strategy, case)

Arguments

kappa

threshold value for the go/no-go decision rule

n2

total sample size for phase II; must be even number

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

Delta1

assumed true treatment effect for standardized difference in means

Delta2

assumed true treatment effect for standardized difference in means

strategy

choose Strategy: 1 ("only best promising"), 2 ("all promising") or 3 (both)

case

different cases: 1 ("nogo"), 21 (treatment 1 is promising, treatment 2 is not), 22 (treatment 2 is promising, treatment 1 is not), 31 (both treatments are promising, treatment 1 is better), 32 (both treatments are promising, treatment 2 is better)

Value

The function Ess_normal() returns the expected sample size for phase III when going to phase III when outcomes are normally distributed and we consider multiarm programs, i.e. several phase III trials with different doses or different treatments are performed

Examples

res <- Ess_normal(kappa = 0.1 ,n2 = 50 ,alpha = 0.05, beta = 0.1,
                            Delta1 = 0.375, Delta2 = 0.625, strategy = 3, case = 31)