Utility based optimal phase II/III drug development planning

The drugdevelopR package enables utility based planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules. The assumed true treatment effects can be assumed fixed (planning is then also possible via user friendly R Shiny App: drugdevelopR) or modelled by a prior distribution. The R Shiny application prior visualizes the prior distributions used in this package. Fast computing is enabled by parallel programming.

Usage

drugdevelopR()

drugdevelopR package and R Shiny App

The drugdevelopR package provides the functions to plan optimal phase II/III drug development programs with

time-to-event endpoint (optimal_tte),

binary endpoint (optimal_binary) and

normally distributed endpoint (optimal_normal),

where the treatment effect is assumed fixed or modelled by a prior. In these settings, optimal phase II/III drug development planning with fixed assumed treatment effects can also be done with the help of the R Shiny application basic. Extensions to the basic setting are

optimal planning of programs including methods for discounting of phase II results (function: optimal_bias, App: bias),

optimal planning of programs with several phase III trials (function: optimal_multitrial, App: multitrial) and

optimal planning of programs with multiple arms (function: optimal_multiarm, App: multiarm).

The R Shiny App drugdevelopR serves as homepage, navigating the different parts of drugdevelopR via links.

References

Kirchner, M., Kieser, M., Goette, H., & Schueler, A. (2016). Utility-based optimization of phase II/III programs. Statistics in Medicine, 35(2), 305-316.

Preussler, S., Kieser, M., and Kirchner, M. (2019). Optimal sample size allocation and go/no-go decision rules for phase II/III programs where several phase III trials are performed. Biometrical Journal, 61(2), 357-378.

Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2019). Optimal designs for phase II/III drug development programs including methods for discounting of phase II results. Submitted to peer-review journal.

Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2019). Optimal designs for multi-arm Phase II/III drug development programs. Submitted to peer-review journal.