drugdevelopR: Utility-Based Optimal Phase II/III Drug Development Planning

Plan optimal sample size allocation and go/no-go decision rules for phase II/III drug development programs with time-to-event, binary or normally distributed endpoints when assuming fixed treatment effects or a prior distribution for the treatment effect, using methods from Kirchner et al. (2016) doi:10.1002/sim.6624 and Preussler (2020). Optimal is in the sense of maximal expected utility, where the utility is a function taking into account the expected cost and benefit of the program. It is possible to extend to more complex settings with bias correction (Preussler S et al. (2020) doi:10.1186/s12874-020-01093-w ), multiple phase III trials (Preussler et al. (2019) doi:10.1002/bimj.201700241 ), multi-arm trials (Preussler et al. (2019) doi:10.1080/19466315.2019.1702092 ), and multiple endpoints (Kieser et al. (2018) doi:10.1002/pst.1861 ).

See also

Author

Maintainer: Lukas D. Sauer sauer@imbi.uni-heidelberg.de (ORCID)

Authors:

Stella Erdmann erdmann@imbi.uni-heidelberg.de

Johannes Cepicka

Marietta Kirchner

Meinhard Kieser