Generic function for optimizing programs with binary endpoints

Generic function for optimizing programs with binary endpoints

Usage

optimal_binary_generic(
  w,
  p0,
  p11,
  p12,
  in1,
  in2,
  n2min,
  n2max,
  stepn2,
  rrgomin,
  rrgomax,
  steprrgo,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 1,
  stepm1 = 0.95,
  stepl1 = 0.85,
  b1,
  b2,
  b3,
  gamma = 0,
  fixed = FALSE,
  num_cl = 1
)

Arguments

w

weight for mixture prior distribution

p0

assumed true rate of control group, see here for details

p11

assumed true rate of treatment group, see here for details

p12

assumed true rate of treatment group, see here for details

in1

amount of information for p11 in terms of sample size, see here for details

in2

amount of information for p12 in terms of sample size, see here for details

n2min

minimal total sample size for phase II; must be an even number

n2max

maximal total sample size for phase II, must be an even number

stepn2

step size for the optimization over n2; must be an even number

rrgomin

minimal threshold value for the go/no-go decision rule

rrgomax

maximal threshold value for the go/no-go decision rule

steprrgo

step size for the optimization over RRgo

alpha

one-sided significance level

beta

type II error rate; i.e. 1 - beta is the power for calculation of the number of events for phase III

c2

variable per-patient cost for phase II in 10^5 $

c3

variable per-patient cost for phase III in 10^5 $

c02

fixed cost for phase II in 10^5 $

c03

fixed cost for phase III in 10^5 $

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

steps1

lower boundary for effect size category "small" in RR scale, default: 1

stepm1

lower boundary for effect size category "medium" in RR scale = upper boundary for effect size category "small" in RR scale, default: 0.95

stepl1

lower boundary for effect size category "large" in RR scale = upper boundary for effect size category "medium" in RR scale, default: 0.85

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

gamma

to model different populations in phase II and III choose gamma != 0, default: 0, see here for details

fixed

choose if true treatment effects are fixed or random, if TRUE p11 is used as fixed effect for p1

num_cl

number of clusters used for parallel computing, default: 1