Generic function for optimizing drug development programs

Generic function for optimizing drug development programs

Usage

optimal_generic(beta, alpha, c2, c3, c02, c03, K, N, S, b1, b2, b3, num_cl)

Arguments

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

num_cl

number of clusters used for parallel computing, default: 1