Optimal phase II/III drug development planning for programs with multiple endpoints

The function optimal_multiple_tte of the drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules (Preussler et. al, 2019). (planning is also possible via user friendly R Shiny App: multiarm). Fast coputing is enabled by parallel programming.

Usage

optimal_multiple_tte(
  hr1,
  hr2,
  id1,
  id2,
  ec,
  d2min,
  d2max,
  stepd2,
  hrgomin,
  hrgomax,
  stephrgo,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  b1,
  b2,
  b3,
  rho,
  fixed,
  num_cl = 1
)

Arguments

hr1

assumed true treatment effect on HR scale for treatment 1

hr2

assumed true treatment effect on HR scale for treatment 2

id1

amount of information for hr1 in terms of number of events

id2

amount of information for hr2 in terms of number of events

ec

control arm event rate for phase II and III

hrgomin

minimal threshold value for the go/no-go decision rule

hrgomax

maximal threshold value for the go/no-go decision rule

stephrgo

stepsize for the optimization over HRgo

alpha

one-sided significance level/ family wise error rate

beta

1-beta (any-pair) power for calculation of the number of events for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

rho

correlation between the two endpoints

fixed

assumed fixed treatment effect

num_cl

number of clusters used for parallel computing, default: 1

n2min

minimal total sample size in phase II, must be divisible by 3

n2max

maximal total sample size in phase II, must be divisible by 3

stepn2

stepsize for the optimization over n2, must be divisible by 3

steps1

lower boundary for effect size category "small" in HR scale, default: 1

stepm1

lower boundary for effect size category "medium" in HR scale = upper boundary for effect size category "small" in HR scale, default: 0.95

stepl1

lower boundary for effect size category "large" in HR scale = upper boundary for effect size category "medium" in HR scale, default: 0.85

Value

The output of the function optimal_multiple_tte is a data.frame containing the optimization results:

Strategy

Strategy, 1: "only best promising" or 2: "all promising"

u

maximal expected utility

HRgo

optimal threshold value for the decision rule to go to phase III

n2

optimal total sample size in phase II

n3

total expected sample size for phase III; rounded to next natural number

n

total sample size in the program; n = n2 + n3

K

maximal costs of the program

pgo

probability to go to phase III

sProg

probability of a successful program

sProg1

probability of a successful program with "small" treatment effect in Phase III

sProg2

probability of a successful program with "medium" treatment effect in Phase III

sProg3

probability of a successful program with "large" treatment effect in Phase III

K2

expected costs for phase II

K3

expected costs for phase III

OP

probability that one endpoint is significant

and further input parameters. res Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure.

drugdevelopR functions

The drugdevelopR package provides the functions

optimal_tte,

optimal_binary and

optimal_normal

to plan optimal phase II/III drug development programs with

time-to-event (treatment effect measured by hazard ratio (HR)),

binary (treatment effect measured by risk ratio (RR)) or

normally distributed (treatment effect measured by standardized difference in means (Delta))

endpoint, where the treatment effect is modelled by a prior. Optimal phase II/III drug development planning with fixed treatment effects can be done with the help of the R Shiny application basic. Extensions are

optimal planning of programs including methods for discounting of phase II results (function: optimal_bias, App: bias),

optimal planning of programs with several phase III trials (function: optimal_multitrial, App: multitrial) and

optimal planning of programs with multiple arms (function: optimal_multiarm, App: multiarm).

References

Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2019). Optimal Designs for Multi-Arm Phase II/III Drug Development Programs. Submitted to peer-review journal.

IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at https://www.iqwig.de/de/methoden/methodenpapier.3020.html, assessed last 15.05.19.

Examples

res <- optimal_multiarm(hr1 = 0.75, hr2 = 0.80, ec = 0.6,# define assumed true HRs and control arm event rate
  id1 = 210, id2 = 420,
  n2min = 30, n2max = 90, stepn2 = 6,                    # define optimization set for n2
  hrgomin = 0.7, hrgomax = 0.9, stephrgo = 0.05,         # define optimization set for HRgo
  alpha = 0.05, beta = 0.1,                              # drug development planning parameters
  c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,               # define fixed and variable costs for phase II and III
  K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  steps1 = 1,                                            # define lower boundary for "small"
  stepm1 = 0.95,                                         # "medium"
  stepl1 = 0.85,                                         # and "large" treatment effect size categories as proposed by IQWiG (2016)
  b1 = 1000, b2 = 2000, b3 = 3000,                       # define expected benefit for a "small", "medium" and "large" treatment effect
  rho = 1/2, fixed = TRUE,                                # correlation and treatment effect
  num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
#> Error in optimal_multiarm(hr1 = 0.75, hr2 = 0.8, ec = 0.6, id1 = 210,     id2 = 420, n2min = 30, n2max = 90, stepn2 = 6, hrgomin = 0.7,     hrgomax = 0.9, stephrgo = 0.05, alpha = 0.05, beta = 0.1,     c2 = 0.75, c3 = 1, c02 = 100, c03 = 150, K = Inf, N = Inf,     S = -Inf, steps1 = 1, stepm1 = 0.95, stepl1 = 0.85, b1 = 1000,     b2 = 2000, b3 = 3000, rho = 1/2, fixed = TRUE, num_cl = 1): unused arguments (id1 = 210, id2 = 420, rho = 1/2, fixed = TRUE)
res
#> Error in eval(expr, envir, enclos): object 'res' not found
cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.
#> Error in comment(res): object 'res' not found