Utility function for multitrial programs deciding between two or three phase III trials for a binary distributed outcome

The utility function calculates the expected utility of our drug development program and is given as gains minus costs and depends on the parameters and the expected probability of a successful program. The utility is in further step maximized by the optimal_multitrial_binary() function.

Usage

utility23_binary(
  n2,
  RRgo,
  w,
  p0,
  p11,
  p12,
  in1,
  in2,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  b1,
  b2,
  b3
)

Arguments

n2

total sample size for phase II; must be even number

RRgo

threshold value for the go/no-go decision rule

w

weight for mixture prior distribution

p0

assumed true rate of control group

p11

assumed true rate of treatment group

p12

assumed true rate of treatment group

in1

amount of information for p11 in terms of sample size

in2

amount of information for p12 in terms of sample size

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

Value

The output of the function utility23_binary() is the expected utility of the program depending on whether two or three phase III trials are performed.

Examples

utility23_binary(n2 = 50, RRgo = 0.8,  w = 0.3, 
                                 alpha = 0.05, beta = 0.1,
                                 p0 = 0.6, p11 =  0.3, p12 = 0.5, 
                                 in1 = 300, in2 = 600, 
                                 c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,
                                 b1 = 1000, b2 = 2000, b3 = 3000)
#>  [1] 740.604220885 216.000000000   0.499091626   0.578956997 389.687099203
#>  [6]  53.551862445   0.024463090   0.192325146   0.253532810  50.000000000
#> [11] 216.000000000   0.063679083  44.000000000   0.006986181  -0.079291991
#> [16]   0.101076391