Utility function for multitrial programs deciding between two or three phase III trials for a normally distributed outcome

The utility function calculates the expected utility of our drug development program and is given as gains minus costs and depends on the parameters and the expected probability of a successful program. The utility is in a further step maximized by the optimal_multitrial_normal() function.

Usage

utility23_normal(
  n2,
  kappa,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  b1,
  b2,
  b3
)

Arguments

n2

total sample size for phase II; must be even number

kappa

threshold value for the go/no-go decision rule

w

weight for mixture prior distribution

Delta1

assumed true treatment effect for standardized difference in means

Delta2

assumed true treatment effect for standardized difference in means

in1

amount of information for Delta1 in terms of sample size

in2

amount of information for Delta2 in terms of sample size

a

lower boundary for the truncation

b

upper boundary for the truncation

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

Value

The output of the function utility23_normal() is the expected utility of the program depending on whether two or three phase III trials are performed.

Examples

utility23_normal(n2 = 50, kappa = 0.2, w = 0.3, alpha = 0.025, beta = 0.1,
                                Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                                a = 0.25, b = 0.75, 
                                c2 = 0.675, c3 = 0.72, c02 = 15, c03 = 20,
                                b1 = 3000, b2 = 8000, b3 = 10000)
#>  [1] 1.844231e+03 3.680000e+02 5.666107e-01 7.885549e-01 2.965022e+02
#>  [6] 1.421218e+01 4.658394e-01 8.456348e-03 3.450019e-06 2.066088e-01
#> [11] 1.400000e+01 0.000000e+00 9.231153e-02 0.000000e+00