Utility function for bias adjustment programs with normally distributed outcomes.

The utility function calculates the expected utility of our drug development program and is given as gains minus costs and depends on the parameters and the expected probability of a successful program. The utility is in a further step maximized by the optimal_bias_normal() function.

Usage

utility_normal_L(
  n2,
  kappa,
  Adj,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K,
  N,
  S,
  steps1,
  stepm1,
  stepl1,
  b1,
  b2,
  b3,
  fixed
)

utility_normal_L2(
  n2,
  kappa,
  Adj,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K,
  N,
  S,
  steps1,
  stepm1,
  stepl1,
  b1,
  b2,
  b3,
  fixed
)

utility_normal_R(
  n2,
  kappa,
  Adj,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K,
  N,
  S,
  steps1,
  stepm1,
  stepl1,
  b1,
  b2,
  b3,
  fixed
)

utility_normal_R2(
  n2,
  kappa,
  Adj,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K,
  N,
  S,
  steps1,
  stepm1,
  stepl1,
  b1,
  b2,
  b3,
  fixed
)

Arguments

n2

total sample size for phase II; must be even number

kappa

threshold value for the go/no-go decision rule

Adj

adjustment parameter

w

weight for mixture prior distribution

Delta1

assumed true treatment effect for standardized difference in means

Delta2

assumed true treatment effect for standardized difference in means

in1

amount of information for Delta1 in terms of sample size

in2

amount of information for Delta2 in terms of sample size

a

lower boundary for the truncation

b

upper boundary for the truncation

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

steps1

lower boundary for effect size category "small", default: 0

stepm1

lower boundary for effect size category "medium" = upper boundary for effect size category "small" default: 0.5

stepl1

lower boundary for effect size category "large" = upper boundary for effect size category "medium", default: 0.8

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

fixed

choose if true treatment effects are fixed or random, if TRUE Delta1 is used as fixed effect

Value

The output of the functions utility_normal_L(), utility_normal_L2(), utility_normal_R() and utility_normal_R2() is the expected utility of the program.

Examples

res <- utility_normal_L(kappa = 0.1, n2 = 50, Adj = 0, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,
                                 K = Inf, N = Inf, S = -Inf, 
                                 steps1 = 0, stepm1 = 0.5, stepl1 = 0.8,
                                 b1 = 3000, b2 = 8000, b3 = 10000, 
                                 fixed = TRUE)
          res <- utility_normal_L2(kappa = 0.1, n2 = 50, Adj = 0, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,
                                 K = Inf, N = Inf, S = -Inf, 
                                 steps1 = 0, stepm1 = 0.5, stepl1 = 0.8,
                                 b1 = 3000, b2 = 8000, b3 = 10000, 
                                 fixed = TRUE)
          res <- utility_normal_R(kappa = 0.1, n2 = 50, Adj = 1, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,
                                 K = Inf, N = Inf, S = -Inf, 
                                 steps1 = 0, stepm1 = 0.5, stepl1 = 0.8,
                                 b1 = 3000, b2 = 8000, b3 = 10000, 
                                 fixed = TRUE)
          res <- utility_normal_R2(kappa = 0.1, n2 = 50, Adj = 1, 
                                 alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, 
                                 in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,
                                 K = Inf, N = Inf, S = -Inf, 
                                 steps1 = 0, stepm1 = 0.5, stepl1 = 0.8,
                                 b1 = 3000, b2 = 8000, b3 = 10000, 
                                 fixed = TRUE)