Expected probability of a successful program deciding between two or three phase III trials for a normally distributed outcome

The function EPsProg23_normal() calculates the expected probability of a successful program with a normally distributed outcome. This function follows a special decision rule in order to determine whether two or three phase III trials should be conducted. First, two phase III trials are performed. Depending on their success, the decision for a third phase III trial is made:

If both trials are successful, no third phase III trial will be conducted.

If only one of the two trials is successful and the other trial has a treatment effect that points in the same direction, a third phase III trial will be conducted with a sample size of N3 = N3(ymin), which depends on an assumed minimal clinical relevant effect (ymin). The third trial then has to be significant at level alpha

If only one of the two trials is successful and the treatment effect of the other points in opposite direction or if none of the two trials are successful, then no third trial is performed and the drug development development program is not successful. In the utility function, this will lead to a utility of -9999.

Usage

EPsProg23_normal(
  kappa,
  n2,
  alpha,
  beta,
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  case,
  size,
  ymin
)

Arguments

kappa

threshold value for the go/no-go decision rule

n2

total sample size for phase II; must be an even number

alpha

significance level

beta

type II error rate; this means that 1 - beta is the power for calculating the sample size for phase III

w

weight for the mixture prior distribution

Delta1

assumed true treatment effect for the standardized difference in means

Delta2

assumed true treatment effect for the standardized difference in means

in1

amount of information for Delta1 in terms of sample size

in2

amount of information for Delta2 in terms of sample size

a

lower boundary for the truncation

b

upper boundary for the truncation

case

number of significant trials needed for approval; possible values are 2 and 3 for this function

size

effect size category; possible values are "small", "medium", "large" and "all"

ymin

assumed minimal clinical relevant effect

Value

The output of the function EPsProg23_normal() is the expected probability of a successful program.

Examples

EPsProg23_normal(kappa = 0.1, n2 = 50, alpha = 0.025, beta = 0.1, w = 0.3,
                                 Delta1 = 0.375, Delta2 = 0.625, in1 = 300, in2 = 600, 
                                 a = 0.25, b = 0.75, 
                                 case = 2, size = "small", ymin = 0.5)
#> [1] 0.5443622