Optimal phase II/III drug development planning when discounting phase II results

The function optimal_bias of the drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules including methods for discounting of phase II results (Preussler et. al, 2020). The assumed true treatment effects can be assumed fixed (planning is then also possible via user friendly R Shiny App: bias) or modelled by a prior distribution. The R Shiny application prior visualizes the prior distributions used in this package. Fast computing is enabled by parallel programming.

Usage

optimal_bias(
  w,
  hr1,
  hr2,
  id1,
  id2,
  d2min,
  d2max,
  stepd2,
  hrgomin,
  hrgomax,
  stephrgo,
  adj = "both",
  lambdamin = NULL,
  lambdamax = NULL,
  steplambda = NULL,
  alphaCImin = NULL,
  alphaCImax = NULL,
  stepalphaCI = NULL,
  alpha,
  beta,
  xi2,
  xi3,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 1,
  stepm1 = 0.95,
  stepl1 = 0.85,
  b1,
  b2,
  b3,
  fixed = FALSE,
  num_cl = 1
)

Arguments

w

weight for mixture prior distribution

hr1

first assumed true treatment effect on HR scale for prior distribution

hr2

second assumed true treatment effect on HR scale for prior distribution

id1

amount of information for hr1 in terms of number of events

id2

amount of information for hr2 in terms of number of events

d2min

minimal number of events for phase II

d2max

maximal number of events for phase II

stepd2

stepsize for the optimization over d2

hrgomin

minimal threshold value for the go/no-go decision rule

hrgomax

maximal threshold value for the go/no-go decision rule

stephrgo

stepsize for the optimization over HRgo

adj

choose type of adjustment: "multiplicative", "additive", "both" (or "all")

lambdamin

minimal adjustment parameter lambda

lambdamax

maximal adjustment parameter lambda

steplambda

stepsize for the adjustment parameter lambda

alphaCImin

minimal alphaCI

alphaCImax

maximal alphaCI

stepalphaCI

stepsize for alphaCI

alpha

one-sided significance level

beta

1-beta power for calculation of the number of events for phase III by Schoenfeld (1981) formula

xi2

event rate for phase II

xi3

event rate for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g., no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g., no constraint

steps1

lower boundary for effect size category "small" in HR scale, default: 1

stepm1

lower boundary for effect size category "medium" in HR scale = upper boundary for effect size category "small" in HR scale, default: 0.95

stepl1

lower boundary for effect size category "large" in HR scale = upper boundary for effect size category "medium" in HR scale, default: 0.85

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

fixed

choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect

num_cl

number of clusters used for parallel computing, default: 1

gamma

to model different populations in phase II and III choose gamma!=0, default: 0

skipII

choose if skipping phase II is an option, default: FASLE

Value

The output of the function optimal_bias is a data.frame containing the optimization results:

Method

Type of adjustment: multipl. (multiplicative) or add. (additive)

u

maximal expected utility

Adj

optimal adjustment parameter (lambda or alphaCI according to Method)

HRgo

optimal threshold value for the decision rule to go to phase III

d2

optimal total number of events for phase II

d3

total expected number of events for phase III; rounded to next natural number

d

total expected number of events in the program; d = d2 + d3

n2

total sample size for phase II; rounded to the next even natural number

n3

total sample size for phase III; rounded to the next even natural number

n

total sample size in the program; n = n2 + n3

K

maximal costs of the program

pgo

probability to go to phase III

sProg

probability of a successful program

sProg1

probability of a successful program with "small" treatment effect in Phase III

sProg2

probability of a successful program with "medium" treatment effect in Phase III

sProg3

probability of a successful program with "large" treatment effect in Phase III

K2

expected costs for phase II

K3

expected costs for phase III

and further input parameters. Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure.

drugdevelopR functions

The drugdevelopR package provides the functions

optimal_tte,

optimal_binary and

optimal_normal

to plan optimal phase II/III drug development programs with

time-to-event (treatment effect measured by hazard ratio (HR)),

binary (treatment effect measured by risk ratio (RR)) or

normally distributed (treatment effect measured by standardized difference in means (Delta))

endpoint, where the treatment effect is modelled by a prior. Optimal phase II/III drug development planning with fixed treatment effects can be done with the help of the R Shiny application basic. Extensions are

optimal planning of programs including methods for discounting of phase II results (function: optimal_bias, App: bias),

optimal planning of programs with several phase III trials (function: optimal_multitrial, App: multitrial) and

optimal planning of programs with multiple arms (function: optimal_multiarm, App: multiarm).

References

IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at https://www.iqwig.de/de/methoden/methodenpapier.3020.html, assessed last 15.05.19.

Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2020). Optimal designs for phase II/III drug development programs including methods for discounting of phase II results. Submitted to peer-review journal.

Schoenfeld, D. (1981). The asymptotic properties of nonparametric tests for comparing survival distributions. Biometrika, 68(1), 316-319.

Examples

res <- optimal_bias(w = 0.3,                                 # define parameters for prior
  hr1 = 0.69, hr2 = 0.88, id1 = 210, id2 = 420,              # (https://web.imbi.uni-heidelberg.de/prior/)
  d2min = 20, d2max = 100, stepd2 = 5,                       # define optimization set for d2
  hrgomin = 0.7, hrgomax = 0.9, stephrgo = 0.05,             # define optimization set for HRgo
  adj = "both",                                              # choose type of adjustment
  lambdamin = 0.2, lambdamax = 1, steplambda = 0.05,         # define optimization set for lambda
  alphaCImin = 0.025, alphaCImax = 0.5, stepalphaCI = 0.025, # define optimization set for alphaCI
  alpha = 0.025, beta = 0.1, xi2 = 0.7, xi3 = 0.7,            # drug development planning parameters
  c2 = 0.75, c3 = 1, c02 = 100, c03 = 150,                   # define fixed and variable costs for phase II and III
  K = Inf, N = Inf, S = -Inf,                                # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  steps1 = 1,                                                # define lower boundary for "small"
  stepm1 = 0.95,                                             # "medium"
  stepl1 = 0.85,                                             # and "large" treatment effect size categories as proposed by IQWiG (2016)
  b1 = 1000, b2 = 2000, b3 = 3000,                           # define expected benefit for a "small", "medium" and "large" treatment effect
  fixed = FALSE,                                             # choose if true treatment effects are fixed or random
  num_cl = 1)                                                # set number of cores used for parallelized computing (check maximum number possible with detectCores())
#> Error in qnorm(1 - alpha): object 'alpha' not found
res
#> Error in eval(expr, envir, enclos): object 'res' not found
cat(comment(res))                                            # displays the optimization sequence, start and finish date of the optimization procedure.
#> Error in comment(res): object 'res' not found