Optimal phase II/III drug development planning for programs with multiple endpoints

The function optimal_multiple_normal of the drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules (Preussler et. al, 2019). (planning is also possible via user friendly R Shiny App: multiarm). Fast coputing is enabled by parallel programming.

Usage

optimal_multiple_normal(
  Delta1,
  Delta2,
  in1,
  in2,
  sigma1,
  sigma2,
  n2min,
  n2max,
  stepn2,
  kappamin,
  kappamax,
  stepkappa,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  steps1 = 0,
  stepm1 = 0.5,
  stepl1 = 0.8,
  b1,
  b2,
  b3,
  rho,
  fixed,
  relaxed = FALSE,
  num_cl = 1
)

Arguments

Delta1

assumed true treatment effect on HR scale for treatment 1

Delta2

assumed true treatment effect on HR scale for treatment 2

in1

amount of information for Delta1 in terms of number of events

in2

amount of information for Delta2 in terms of number of events

sigma1

variance of endpoint 1

sigma2

variance of endpoint 1

n2min

minimal total sample size in phase II, must be divisible by 3

n2max

maximal total sample size in phase II, must be divisible by 3

stepn2

stepsize for the optimization over n2, must be divisible by 3

kappamin

minimal threshold value for the go/no-go decision rule

kappamax

maximal threshold value for the go/no-go decision rule

stepkappa

stepsize for the optimization over HRgo

alpha

one-sided significance level/ family wise error rate

beta

1-beta (any-pair) power for calculation of the number of events for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

steps1

lower boundary for effect size category "small" in HR scale, default: 1

stepm1

lower boundary for effect size category "medium" in HR scale = upper boundary for effect size category "small" in HR scale, default: 0.95

stepl1

lower boundary for effect size category "large" in HR scale = upper boundary for effect size category "medium" in HR scale, default: 0.85

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

rho

correlation between the two endpoints

fixed

assumed fixed treatment effect

relaxed

relaxed or strict decision rule

num_cl

number of clusters used for parallel computing, default: 1

Value

The output of the function optimal_multiple_tte is a data.frame containing the optimization results:

u

maximal expected utility

Kappa

optimal threshold value for the decision rule to go to phase III

n2

total sample size for phase II

n3

total sample size for phase III; rounded to the next even natural number

n

total sample size in the program; n = n2 + n3

K

maximal costs of the program

pgo

probability to go to phase III

sProg

probability of a successful program

sProg1

probability of a successful program with "small" treatment effect in Phase III

sProg2

probability of a successful program with "medium" treatment effect in Phase III

sProg3

probability of a successful program with "large" treatment effect in Phase III

K2

expected costs for phase II

K3

expected costs for phase III

and further input parameters.

Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure. @examples #res <- optimal_multiple_normal(Delta1 = 0.75, Delta2 = 0.80, # define assumed true HRs and control arm event ratein1=300, in2=600, sigma1 = 8, sigma2= 12,n2min = 30, n2max = 90, stepn2 = 6, # define optimization set for n2kappamin = 0.7, kappamax = 0.9, stepkappa = 0.05, # define optimization set for HRgoalpha = 0.05, beta = 0.1, # drug development planning parametersc2 = 0.75, c3 = 1, c02 = 100, c03 = 150, # define fixed and variable costs for phase II and IIIK = Inf, N = Inf, S = -Inf, # set maximal costs/ expected sample size for the program or minimal expected probability of a successful programsteps1 = 0, # define lower boundary for "small"stepm1 = 0.5, # "medium"stepl1 = 0.88, # and "large" treatment effect size categories as proposed by IQWiG (2016)b1 = 1000, b2 = 2000, b3 = 3000, # define expected benefit for a "small", "medium" and "large" treatment effectrho = 0.5, relaxed = TRUE, # relaxed "TRUE"fixed = TRUE, # treatment effectnum_cl = 1) # set number of cores used for parallelized computing (check maximum number possible with detectCores())

#res #cat(comment(res)) # displays the optimization sequence, start and finish date of the optimization procedure.

drugdevelopR functions

The drugdevelopR package provides the functions

optimal_tte,

optimal_binary and

optimal_normal

to plan optimal phase II/III drug development programs with

time-to-event (treatment effect measured by hazard ratio (HR)),

binary (treatment effect measured by risk ratio (RR)) or

normally distributed (treatment effect measured by standardized difference in means (Delta))

endpoint, where the treatment effect is modelled by a prior. Optimal phase II/III drug development planning with fixed treatment effects can be done with the help of the R Shiny application basic. Extensions are

optimal planning of programs including methods for discounting of phase II results (function: optimal_bias, App: bias),

optimal planning of programs with several phase III trials (function: optimal_multitrial, App: multitrial) and

optimal planning of programs with multiple arms (function: optimal_multiarm, App: multiarm).

References

Preussler, S., Kirchner, M., Goette, H., Kieser, M. (2019). Optimal Designs for Multi-Arm Phase II/III Drug Development Programs. Submitted to peer-review journal.

IQWiG (2016). Allgemeine Methoden. Version 5.0, 10.07.2016, Technical Report. Available at https://www.iqwig.de/de/methoden/methodenpapier.3020.html, assessed last 15.05.19.

Last Edited By

Johannes Cepicka

Last Edit Date

2022-04-23