Optimal phase II/III drug development planning where several phase III trials are performed

The drugdevelopR package enables planning of phase II/III drug development programs with optimal sample size allocation and go/no-go decision rules. For normally distributed endpoints the treatment effect is measured by the standardized difference in means (Delta). The assumed true treatment effects can be assumed fixed or modelled by a prior distribution. The R Shiny application prior visualizes the prior distributions used in this package. Fast coputing is enabled by parallel programming.

Usage

optimal_multitrial_normal(
  w,
  Delta1,
  Delta2,
  in1,
  in2,
  a,
  b,
  n2min,
  n2max,
  stepn2,
  kappamin,
  kappamax,
  stepkappa,
  alpha,
  beta,
  c2,
  c3,
  c02,
  c03,
  K = Inf,
  N = Inf,
  S = -Inf,
  b1,
  b2,
  b3,
  case,
  strategy = TRUE,
  fixed = FALSE,
  num_cl = 1
)

Arguments

w

weight for mixture prior distribution

Delta1

assumed true treatment effect for standardized difference in means

Delta2

assumed true treatment effect for standardized difference in means

in1

amount of information for Delta2 in terms of sample size

in2

amount of information for Delta1 in terms of sample size

a

lower boundary for the truncation

b

upper boundary for the truncation

n2min

minimal total sample size for phase II; must be even number

n2max

maximal total sample size for phase II, must be even number

stepn2

stepsize for the optimization over n2; must be even number

kappamin

minimal threshold value for the go/no-go decision rule

kappamax

maximal threshold value for the go/no-go decision rule

stepkappa

stepsize for the optimization over kappa

alpha

significance level

beta

1-beta power for calculation of sample size for phase III

c2

variable per-patient cost for phase II

c3

variable per-patient cost for phase III

c02

fixed cost for phase II

c03

fixed cost for phase III

K

constraint on the costs of the program, default: Inf, e.g. no constraint

N

constraint on the total expected sample size of the program, default: Inf, e.g. no constraint

S

constraint on the expected probability of a successful program, default: -Inf, e.g. no constraint

b1

expected gain for effect size category "small"

b2

expected gain for effect size category "medium"

b3

expected gain for effect size category "large"

case

choose case: "at least 1, 2 or 3 significant trials needed for approval"

strategy

choose strategy: "conduct 1, 2, 3 or 4 trial"; TRUE calculates all strategies of the selected Case

fixed

choose if true treatment effects are fixed or random, if TRUE hr1 is used as fixed effect

num_cl

number of clusters used for parallel computing, default: 1

steps1

lower boundary for effect size category "small", default: 0

stepm1

lower boundary for effect size category "medium" = upper boundary for effect size category "small" default: 0.5

stepl1

lower boundary for effect size category "large" = upper boundary for effect size category "medium", default: 0.8

Value

The output of the function optimal_multitrial_normal is a data.frame containing the optimization results:

Case

Case: "number of significant trials needed"

Strategy

Strategy: "number of conducted trials"

u

maximal expected utility

kappa

optimal threshold value for the decision rule to go to phase III

n2

total sample size for phase II

n3

total sample size for phase III; rounded to the next even natural number

n

total sample size in the program; n = n2 + n3

K

maximal costs of the program

pgo

probability to go to phase III

sProg

probability of a successful program

sProg1

probability of a successful program with "small" treatment effect in Phase III

sProg2

probability of a successful program with "medium" treatment effect in Phase III

sProg3

probability of a successful program with "large" treatment effect in Phase III

K2

expected costs for phase II

K3

expected costs for phase III

and further input parameters. Taking cat(comment()) of the data.frame object lists the used optimization sequences, start and finish date of the optimization procedure.

drugdevelopR functions

The drugdevelopR package provides the functions

optimal_tte,

optimal_binary or

optimal_normal

to plan optimal phase II/III drug development programs with

time-to-event (treatment effect measured by hazard ratio (HR)),

binary (treatment effect measured by risk ratio (RR)) and

normally distributed (treatment effect measured by standardized difference in means (Delta))

endpoint, where the treatment effect is assumed fixed or modelled by a prior. Optimal phase II/III drug development planning with fixed treatment effects can also be done with the help of the R Shiny application basic. Extensions are

optimal planning of programs including methods for discounting of phase II results (function: optimal_bias, App: bias),

optimal planning of programs with several phase III trials (function: optimal_multitrial, App: multitrial) and

optimal planning of programs with multiple arms (function: optimal_multiarm, App: multiarm).

References

Cohen, J. (1988). Statistical power analysis for the behavioral sciences.

Examples

res <- optimal_multitrial_normal(w=0.3,                             # define parameters for prior
  Delta1 = 0.375, Delta2 = 0.625, in1=300, in2=600,      # (https://web.imbi.uni-heidelberg.de/prior/)
  a = 0.25, b = 0.75,
  n2min = 20, n2max = 100, stepn2 = 4,                   # define optimization set for n2
  kappamin = 0.02, kappamax = 0.2, stepkappa = 0.02,     # define optimization set for kappa
  alpha = 0.05, beta = 0.1,                              # drug development planning parameters
  c2 = 0.675, c3 = 0.72, c02 = 15, c03 = 20,             # define fixed and variable costs for phase II and III
  K = Inf, N = Inf, S = -Inf,                            # set maximal costs/ expected sample size for the program or minimal expected probability of a successful program
  b1 = 3000, b2 = 8000, b3 = 10000,                      # define expected benefit for a "small", "medium" and "large" treatment effect                                             # assume different/same population structures in phase II and III
  case = 1, strategy = TRUE,                             # chose Case and Strategy
  fixed = TRUE,                                          # choose if true treatment effects are fixed or random
  num_cl = 1)                                            # set number of cores used for parallelized computing (check maximum number possible with detectCores())
#> Error in qnorm(1 - alpha): object 'alpha' not found
res
#> Error in eval(expr, envir, enclos): object 'res' not found

cat(comment(res))                                        # displays the optimization sequence, start and finish date of the optimization procedure.
#> Error in comment(res): object 'res' not found